Feat/collection flow

+++ src/app/api/media/[id]/route.ts
@@ -1,8 +1,11 @@
+import { getUserId } from 'core/getUserId';
 import { NextResponse } from 'next/server';
 import { deleteMediaCollectionUseCase } from 'server/modules/media/deleteMediaCollection';
 
+export const dynamic = 'force-dynamic';
+
 export async function DELETE(req: Request, ctx: any) {
-  const { id } = ctx.params;
+  const { id, collectionId } = ctx.params;
 
   if (!id) {
     return NextResponse.json({
@@ -10,8 +13,18 @@ export async function DELETE(req: Request, ctx: any) {
     }, { status: 400 });
   }
 
+  const userId = await getUserId();
+
+  if (!userId) {
+    return NextResponse.json({
+      error: 'Not authenticate',
+    }, { status: 400 });
+  }
+
   await deleteMediaCollectionUseCase({
     mediaId: id,
+    userId,
+    collectionId,
   });
 
   return NextResponse.json({}, { status: 201 });
